+++ test_macro.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env argorator
-
-# INPUT_FILE (file): Input file to process
-
-echo "Testing iteration macros"
-
-# for line in $INPUT_FILE
-echo "Processing: $line"